resources/requirements.txt
qrcode[pil]>=7.4
Pillow>=10.0
opencv-python-headless>=4.8
